Moved doc viewer logic for generating search index, queries, etc into a web worker.
