+++ packages/altair-app/package.json
@@ -50,6 +50,7 @@
     "body-parser": "^1.19.0",
     "codemirror": "^5.48.0",
     "codemirror-graphql": "^0.11.1",
+    "comlink": "^4.2.0",
     "comment-regex": "^1.0.1",
     "cookie-parser": "^1.4.4",
     "core-js": "^3.1.4",
