+++ packages/altair-app/src/app/components/doc-viewer/doc-viewer/doc-viewer.component.ts
@@ -10,7 +10,7 @@ import {
   OnDestroy
 } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
-import { of } from 'rxjs';
+import { of, from } from 'rxjs';
 
 import { AltairConfig } from '../../../config';
 import { debug } from 'app/utils/logger';
@@ -27,6 +27,11 @@ import {
 } from 'graphql';
 import { DocumentIndexEntry } from '../models';
 import { fadeInOutAnimationTrigger } from 'app/animations';
+import * as Comlink from 'comlink';
+import { GqlService } from 'app/services';
+import getRootTypes from 'app/utils/get-root-types';
+
+const DocUtils: any = Comlink.wrap(new Worker('../doc-utils.worker', { type: 'module' }));
 
 @Component({
   selector: 'app-doc-viewer',
@@ -73,8 +78,11 @@ export class DocViewerComponent implements OnChanges, OnDestroy {
   searchResult: DocumentIndexEntry[] = [];
   searchTerm = '';
 
+  docUtilWorker: any;
+
   constructor(
     private translate: TranslateService,
+    private gqlService: GqlService,
     private _sanitizer: DomSanitizer,
     private altairConfig: AltairConfig,
   ) {
@@ -94,16 +102,17 @@ export class DocViewerComponent implements OnChanges, OnDestroy {
     }
   }
 
-  updateDocs(schema: GraphQLSchema) {
+  async updateDocs(schema: GraphQLSchema) {
     debug.log(schema);
-    this.rootTypes = [
-      schema.getQueryType(),
-      schema.getMutationType(),
-      schema.getSubscriptionType()
-    ].filter(Boolean) as GraphQLObjectType[];
+    this.rootTypes = getRootTypes(schema);
 
     try {
-      this.generateIndex(schema);
+      const docUtilWorker = await this.getDocUtilsWorker();
+      const sdl = await this.gqlService.getSDL(schema);
+      await docUtilWorker.updateSchema(sdl);
+      // this.generateIndex(schema);
+      this.index = await this.docUtilWorker.generateSearchIndex();
+      debug.log('Worker index:', this.index);
       this.hasSearchIndex = true;
     } catch (err) {
       debug.log('Error while generating index.', err);
@@ -111,129 +120,7 @@ export class DocViewerComponent implements OnChanges, OnDestroy {
     }
   }
 
-  /**
-   * Generate the search index from the schema
-   * @param schema
-   */
-  generateIndex(schema: GraphQLSchema) {
-    let getFieldsIndices: (
-      fields: GraphQLFieldMap<any, any>,
-      type: GraphQLType,
-      isQuery: Boolean,
-      curIndexStack: DocumentIndexEntry[]
-    ) => DocumentIndexEntry[];
-    let getTypeIndices: (type: GraphQLObjectType, isRoot: boolean, curIndexStack: DocumentIndexEntry[]) => DocumentIndexEntry[];
-
-    /**
-     * Gets the indices for fields
-     */
-    getFieldsIndices = (fields, type, isQuery: boolean, curIndexStack) => {
-      let index: DocumentIndexEntry[] = [];
-
-      Object.keys(fields).forEach((fieldKey: any) => {
-        const field = fields[fieldKey];
-
-        // For each field, create an entry in the index
-        const fieldIndex = {
-          search: field.name,
-          name: field.name,
-          description: field.description ? field.description : '',
-          args: field.args,
-          cat: 'field',
-          type: (type as GraphQLObjectType).name,
-          isQuery,
-          highlight: 'field'
-        };
-        index = [ ...index, fieldIndex ];
-
-        // For each argument of the field, create an entry in the index for the field,
-        // searchable by the argument name
-        if (field.args && field.args.length) {
-          field.args.forEach(arg => {
-            index = [...index, {
-              ...fieldIndex,
-              search: arg.name,
-              highlight: 'argument'
-            }];
-          });
-        }
-
-        // If the field has a type, get indices for the type as well
-        if (field.type) {
-          index = [
-            ...index,
-            ...getTypeIndices((field.type as GraphQLObjectType), false, [...curIndexStack, ...index]).filter(val => !!val),
-          ];
-        }
-
-      });
-
-      return index;
-    };
-
-    /**
-     * Gets the indices for types
-     * @param  {object} type the type object
-     * @param  {boolean} isRoot specifies if the type is a root level type
-     * @param  {array} curIndexStack contains all the currently mapped indices in the stack
-     * @return {array}            the indices for the given type
-     */
-    getTypeIndices = (type, isRoot, curIndexStack) => {
-      let fields = null;
-
-      // If a type does not have a name, don't process it
-      if (!(type as GraphQLObjectType).name) {
-        return [];
-      }
-
-      // If any type is already in the index, then don't process the type again
-      if (curIndexStack.some(x => x.name === type.name && x.cat === 'type')) {
-        return [];
-      }
-
-      if (type.getFields) {
-        fields = type.getFields();
-      }
-
-      const index = [
-        {
-          search: type.name,
-          name: type.name,
-          cat: 'type',
-          description: type.description ? type.description : '',
-          isRoot,
-          highlight: 'type'
-        }
-      ];
-
-      if (fields) {
-        return [
-          ...index,
-          ...getFieldsIndices(fields, type, isRoot, [ ...curIndexStack, ...index ]).filter(val => !!val)];
-      }
-
-      return index;
-    };
-
-    this.index = [];
-
-    // Store the indices of all the types and fields
-    this.rootTypes.forEach(type => {
-      this.index = [...this.index, ...getTypeIndices(type, true, this.index)];
-    });
-
-    // Get types from typeMap into index as well, excluding the __
-    const schemaTypeMap = schema.getTypeMap();
-    Object.keys(schemaTypeMap).forEach(key => {
-      if (!/^__/.test(key)) {
-        this.index = [...this.index, ...getTypeIndices(schemaTypeMap[key] as GraphQLObjectType, false, this.index)];
-      }
-    });
-
-    debug.log('Index: ', this.index);
-  }
-
-  autocompleteSource = (term: string) => of(this.index.filter(item => new RegExp(term, 'i').test(item.search)));
+  autocompleteSource = (term: string) => from(this.getDocUtilsWorker().then(_ => _.searchDocs(term)));
   autocompleteListFormatter = (data: DocumentIndexEntry) => {
     const html = `
       <div class='doc-viewer-autocomplete-item'>
@@ -255,7 +142,7 @@ export class DocViewerComponent implements OnChanges, OnDestroy {
   /**
    * search through the docs for the provided term
    */
-  searchDocs(term: string | { name: string }) {
+  async searchDocs(term: string | { name: string }) {
     if (typeof term !== 'string') {
       term = term.name;
     }
@@ -265,7 +152,8 @@ export class DocViewerComponent implements OnChanges, OnDestroy {
     }
     this.updateDocHistory();
     this.setDocViewChange.next({ view: 'search' });
-    this.searchResult = this.index.filter(item => new RegExp(term as string, 'i').test(item.search));
+    const docUtilWorker = await this.getDocUtilsWorker();
+    this.searchResult = await docUtilWorker.searchDocs(term);
     debug.log(this.searchResult);
   }
 
@@ -322,156 +210,15 @@ export class DocViewerComponent implements OnChanges, OnDestroy {
     this.setDocViewChange.next({ view: 'field', name: name.replace(/[\[\]!]/g, ''), parentType: parentType.replace(/[\[\]!]/g, '') });
   }
 
-  /**
-   * Generate the query for the specified field
-   * @param name name of the current field
-   * @param parentType parent type of the current field
-   * @param parentFields preceding parent field and type combinations
-   */
-  generateQuery(name: string, parentType: string) {
-    let query = '';
-    let hasArgs = false;
-
-    if (!this.gqlSchema) {
-      return;
-    }
-
-    // Add the root type of the query
-    switch (parentType) {
-      case this.gqlSchema.getQueryType() && this.gqlSchema.getQueryType()!.name:
-        query += 'query';
-        break;
-      case this.gqlSchema.getMutationType() && this.gqlSchema.getMutationType()!.name:
-        query += 'mutation';
-        break;
-      case this.gqlSchema.getSubscriptionType() && this.gqlSchema.getSubscriptionType()!.name:
-        query += 'subscription';
-        break;
-      default:
-        query += `fragment _____ on ${parentType}`;
-        hasArgs = true;
-    }
-
-    const fieldData = this.generateFieldData(name, parentType, [], 1);
-
-    if (!fieldData) {
-      return;
-    }
-
-    // Add the query fields
-    query += `{\n${fieldData.query}\n}`;
-
-    const meta = { ...fieldData.meta };
-
-    // Update hasArgs option
-    meta.hasArgs = hasArgs || meta.hasArgs;
-
-    return { query, meta };
-  }
-
-  /**
-   * Generate the query for the specified field
-   * @param name name of the current field
-   * @param parentType parent type of the current field
-   * @param parentFields preceding parent field and type combinations
-   * @param level current depth level of the current field
-   */
-  private generateFieldData(
-    name: string,
-    parentType: string,
-    parentFields: { name: string, type: string }[],
-    level: number
-  ): { query: string, meta: { hasArgs?: boolean } } {
-
-    if (!name || !parentType || !parentFields) {
-      return { query: '', meta: {} };
-    }
-    const tabSize = this.tabSize || 2;
-    const parentTypeObject = this.gqlSchema.getType(parentType) as GraphQLObjectType | undefined;
-    const field = parentTypeObject && parentTypeObject.getFields()[name];
-
-    if (!field) {
-      return { query: '', meta: {} };
-    }
-    const meta = {
-      hasArgs: false
-    };
-
-    // Start the query with the field name
-    let fieldStr: string = ' '.repeat(level * tabSize) + field.name;
-
-    // If the field has arguments, add them
-    if (field.args && field.args.length) {
-      meta.hasArgs = true;
-
-      const argsList = field.args.reduce((acc, cur) => {
-        return acc + ', ' + cur.name + ': ______';
-      }, '').substring(2);
-
-      fieldStr += `(${argsList})`;
-    }
-
-    // Retrieve the current field type
-    const curTypeName = this.cleanName(field.type.inspect());
-    const curType = this.gqlSchema.getType(curTypeName) as GraphQLObjectType | undefined;
-
-    // Don't add a field if it has been added in the query already.
-    // This happens when there is a recursive field
-    if (parentFields.filter(x => x.type === curTypeName).length) {
-      return { query: '', meta: {} };
-    }
-
-    // Stop adding new fields once the specified level depth limit is reached
-    if (level >= this.addQueryDepthLimit) {
-      return { query: '', meta: {} };
-    }
-
-    // Get all the fields of the field type, if available
-    const innerFields = curType && curType.getFields && curType.getFields();
-    let innerFieldsData = '';
-    if (innerFields) {
-      innerFieldsData = Object.keys(innerFields).reduce((acc, cur) => {
-        // Don't add a field if it has been added in the query already.
-        // This happens when there is a recursive field
-        if (parentFields.filter(x => x.name === cur && x.type === curTypeName).length) {
-          return '';
-        }
-
-        const curInnerFieldData = this.generateFieldData(cur, curTypeName, [...parentFields, { name, type: curTypeName }], level + 1);
-        if (!curInnerFieldData) {
-          return acc;
-        }
-
-        const curInnerFieldStr: String = curInnerFieldData.query;
-
-        // Set the hasArgs meta if the inner field has args
-        meta.hasArgs = meta.hasArgs || curInnerFieldData.meta.hasArgs || false;
-
-        // Don't bother adding the field if there was nothing generated.
-        // This should fix the empty line issue in the inserted queries
-        if (!curInnerFieldStr) {
-          return acc;
-        }
-
-        // Join all the fields together
-        return acc + '\n' + curInnerFieldStr;
-      }, '').substring(1);
-    }
-
-    // Add the inner fields with braces if available
-    if (innerFieldsData) {
-      fieldStr += `{\n${innerFieldsData}\n`;
-      fieldStr += ' '.repeat(level * tabSize) + `}`;
-    }
-
-    return { query: fieldStr, meta };
-  }
-
-  addToEditor(name: string, parentType: string) {
+  async addToEditor(name: string, parentType: string) {
     if (!this.hasSearchIndex) {
       return false;
     }
-    const generatedQuery = this.generateQuery(name, parentType);
+    const docUtilsWorker = await this.getDocUtilsWorker();
+    const generatedQuery = await docUtilsWorker.generateQuery(name, parentType, {
+      tabSize: this.tabSize,
+      addQueryDepthLimit: this.addQueryDepthLimit,
+    });
     if (generatedQuery) {
       this.addQueryToEditorChange.next(generatedQuery);
     }
@@ -479,6 +226,13 @@ export class DocViewerComponent implements OnChanges, OnDestroy {
 
   exportSDL() {
     this.exportSDLChange.next();
+  }
+
+  async getDocUtilsWorker() {
+    if (!this.docUtilWorker) {
+      this.docUtilWorker = await new DocUtils();
+    }
+    return this.docUtilWorker;
   }
 
   onResize(resizeFactor: number) {
