+++ packages/altair-app/src/app/utils/get-root-types.ts
@@ -1,0 +1,9 @@
+import { GraphQLSchema, GraphQLObjectType } from 'graphql';
+
+export default (schema: GraphQLSchema) => {
+  return [
+    schema.getQueryType(),
+    schema.getMutationType(),
+    schema.getSubscriptionType()
+  ].filter(Boolean) as GraphQLObjectType[];
+}
