+++ packages/altair-app/src/scss/components/_doc-viewer.scss
