+++ packages/altair-app/src/scss/components/_url-box.scss
