+++ packages/altair-app/src/tsconfig.app.json
@@ -8,7 +8,8 @@
   },
   "exclude": [
     "test.ts",
-    "**/*.spec.ts"
+    "**/*.spec.ts",
+    "**/*.worker.ts"
   ],
   "angularCompilerOptions": {
     "preserveWhitespaces": true
